AI-ML-homework/senior-ocr-pwa/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pytesseract==0.3.10
opencv-python-headless==4.9.0.80
Pillow==10.2.0
numpy==1.26.3
python-dotenv==1.0.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
